+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Tue, 09 Aug 2022 22:19:14 GMT and should not be manually modified.
+This log was last generated on Wed, 10 Aug 2022 17:05:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 8.2.6
+
+Wed, 10 Aug 2022 17:05:04 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v11.9.0
 
 ## 8.2.5